==> Code/00_Gathering_tweets.ipynb <==
import json
import tweepy
import pandas as pd
import numpy as np
import datetime
import time

# Research and code usage
- [Twitter Tutorial](http://socialmedia-class.org/twittertutorial.html)
- [Code for 'tweets_to_dataframe' function](https://www.youtube.com/watch?v=WX0MDddgpA4&list=PL5tcWHG-UPH2zBfOz40HSzcGUPAVOOnu1&index=3)
- [Tweepy doc](http://docs.tweepy.org/en/v3.8.0/api.html)

ACCESS_TOKEN = # ACCESS TOKEN (removed for privacy)
ACCESS_SECRET = # ACCESS SECRET (removed for privacy)
CONSUMER_KEY = # CONSUMER KEY (removed for privacy)
CONSUMER_SECRET = #CONSUMER SECRET (removed for privacy)

auth = tweepy.OAuthHandler(CONSUMER_KEY, CONSUMER_SECRET)
auth.set_access_token(ACCESS_TOKEN, ACCESS_SECRET)
api = tweepy.API(auth)

# https://www.youtube.com/watch?v=WX0MDddgpA4&list=PL5tcWHG-UPH2zBfOz40HSzcGUPAVOOnu1&index=3
# Also got the help of Shreya Shenoy
def tweets_to_dataframe(tweets):
    df = pd.DataFrame(data=[tweet.text for tweet in tweets], columns=['tweets'])
#     df['len'] = np.array([len(tweet.text) for tweet in tweets])
    df['id'] = np.array([tweet.id for tweet in tweets])
    df['name'] = np.array([tweet.author.name for tweet in tweets])
    df['location'] = np.array([tweet.user.location for tweet in tweets])
    df['coordinates'] = np.array([tweet.coordinates for tweet in tweets])
    df['created_at'] = np.array([tweet.created_at for tweet in tweets])
#     df['entities'] = np.array([tweet.entities for tweet in tweets])
#     df['favorite'] = np.array([tweet.favorite for tweet in tweets])
    df['favorite_count'] = np.array([tweet.favorite_count for tweet in tweets])
#     df['retweets'] = np.array([tweet.retweets for tweet in tweets])
    df['geo'] = np.array([tweet.geo for tweet in tweets])
#     df['geo'] = np.array([tweet.geo.coordinates if tweet.geo != None else 'None' for tweet in tweets])
    df['place'] = np.array([tweet.place for tweet in tweets])
    df['source'] = np.array([tweet.source for tweet in tweets])

    return df

def grabbing_tweets(query, date, num_iter):
    df_list = []
    current_date = date
    
    for _ in range(num_iter):
        raw_json = api.search(q=query, count=100, until=current_date)
        df = tweets_to_dataframe(raw_json)
        df['created_at'] = df['created_at'].apply(datetime.datetime.date)
        df_list.append(df)
        current_date = df['created_at'].min()
        time.sleep(60)
    return pd.concat(df_list, axis=0)

df_clorox = grabbing_tweets('#clorox', '2020-05-13', 15)

df_clorox.shape

df_clorox

df_clorox.loc[~df_clorox['coordinates'].isnull()]

df_clorox.shape

df_clorox_2 = grabbing_tweets('clorox', '2020-05-13', 10)

df_clorox_2.shape

df_clorox_2.loc[~df_clorox_2['coordinates'].isnull()]

df_clorox_2.tail()

df_clorox_wipes = grabbing_tweets('#cloroxwipes', '2020-05-13', 15)

df_clorox_wipes.tail()

df_clorox_wipes.shape

df_cloroxwipes_nohash = grabbing_tweets('cloroxwipes', '2020-05-13', 15)

df_cloroxwipes_nohash.shape

df_cloroxwipes_nohash.head()

df_lysol = grabbing_tweets('#lysol', '2020-05-13', 15)

df_lysol.shape

df_lysol

df_hydrogen_peroxide = grabbing_tweets('hydrogenperoxide', '2020-05-13', 15)

df_hydrogen_peroxide.shape

df_hydrogen_peroxide.head()

df_hydrogen_peroxide['tweets'][3]

df_facemask = grabbing_tweets('facemask', '2020-05-13', 15)

df_facemask.shape

df_facemask.tail()

df_isopropyl_alcohol = grabbing_tweets('isopropyl alcohol', '2020-05-13', 15)

df_isopropyl_alcohol.shape

df_isopropyl_alcohol.tail()

#papertowels
df_papertowels = grabbing_tweets('#papertowels', '2020-05-13', 15)

df_papertowels.shape

df_papertowels.tail()

hand_sanitizer = grabbing_tweets('#handsanitizer', '2020-05-13', 10)

hand_sanitizer.shape

hand_sanitizer.tail()

disinfectant = grabbing_tweets('disinfectant', '2020-05-13', 10)

disinfectant.shape

disinfectant.tail()

toilet_paper = grabbing_tweets('toiletpaper', '2020-05-13', 10)

toilet_paper.shape

toilet_paper.tail()

all_commodities = pd.concat([df_clorox,
            df_clorox_2,
            df_clorox_wipes,
            df_cloroxwipes_nohash,
            df_lysol,
            df_hydrogen_peroxide,
            df_facemask, 
            df_isopropyl_alcohol,
            df_papertowels, 
            hand_sanitizer,
            disinfectant,
            toilet_paper], axis = 0)

all_commodities.shape

bool_series = all_commodities['id'].duplicated(keep=False)
all_commodities = all_commodities[~bool_series]
all_commodities.shape

all_commodities.to_csv('datasets/all_commodities_tweets.csv', index=False)
