==> Code/01_Gathering_coordinates.ipynb <==
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
import json
import datetime
from pandas.io.json import json_normalize

all_tweets = pd.read_csv('datasets/all_commodities_tweets.csv')

all_tweets

tweets_with_geo = all_tweets.loc[~all_tweets['geo'].isnull()]

tweets_with_geo['geo'].value_counts()

tweets_with_geo.count()

tweets_with_geo.loc[tweets_with_geo['geo'] == "{'type': 'Point', 'coordinates': [40.7142, -74.0064]}", :]

tweets_with_geo.loc[tweets_with_geo['geo'] == "{'type': 'Point', 'coordinates': [40.7352, -74.1849]}", :]

tweets_with_geo['tweets']

tweets_with_geo.iloc[1]['name']

tweets_with_geo.iloc[6]['coordinates']

tweets_with_geo.iloc[19]['coordinates']

tweets_with_geo.iloc[21]['tweets']

tweets_with_geo.iloc[23]['tweets']

tweets_with_geo.iloc[26]['tweets']

tweets_with_geo.iloc[27]['tweets']

tweets_with_geo.iloc[28]['tweets']

tweets_with_geo.iloc[29]['coordinates']

tweets_with_geo.iloc[34]['coordinates']

tweets_with_geo.iloc[38]['coordinates']

tweets_with_geo.iloc[47]['coordinates']

tweets_with_geo.iloc[50]['tweets']
