==> Code/02_Cleaning_tweets.ipynb <==
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup
from nltk.tokenize import RegexpTokenizer, word_tokenize
import pandas as pd
import regex as re

from nltk.corpus import stopwords 
from bs4 import BeautifulSoup  

all_tweets = pd.read_csv('datasets/all_commodities_tweets.csv')

all_tweets

# This function has been taken from Matt Brems NLP I lecture. Very little has been changed.
# Only difference is a change in the function name
def tweets_to_words(raw_text):
    
    text = BeautifulSoup(raw_text).get_text()
    
    letters_only = re.sub("[^a-zA-Z]", " ", text)
    
    words = letters_only.lower().split()
    
    stops = set(stopwords.words('english'))
    
    meaningful_words = [w for w in words if w not in stops]

    return(" ".join(meaningful_words))

all_tweets['tweets'][2]

tweets_to_words(all_tweets['tweets'][2])

total_tweets = all_tweets.shape[0]

# This code is heavily borrowed from NLP I lesson by Matt Brems.
# There have been some slight modifications made to fit this project.
# Initialize an empty list to hold the clean titles.
clean_tweets = []

print("Cleaning and parsing the tweets...")

# Instantiate counter.
j = 0

# Looping through tweets...
for tweet in all_tweets['tweets']:
    
    # Convert title to words, then append to clean_titles.
    clean_tweets.append(tweets_to_words(tweet))
    
    # If the index is divisible by 1000, print a message.
    if (j + 1) % 1000 == 0:
        print(f'Tweets {j + 1} of {total_tweets}.')
    
    j += 1

clean_tweets_df = pd.DataFrame(clean_tweets, columns=['clean_tweets'])

clean_tweets_df.head()

all_tweets = pd.concat([all_tweets, clean_tweets_df], axis=1)

all_tweets.to_csv('datasets/cleaned_tweets.csv', index=False)
