==> Code/03_Preprocessing_Modeling.ipynb <==
import pandas as pd

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.cluster import DBSCAN
from sklearn.metrics import silhouette_score
from sklearn.preprocessing import StandardScaler

tweets = pd.read_csv('datasets/cleaned_tweets.csv')

tweets.head()

tweets.dtypes

tweets.isnull().sum()

tweets.dropna(subset = ['clean_tweets'] , inplace=True)

tweets.isnull().sum()

cvec = CountVectorizer(stop_words=None , ngram_range=(1, 2))

cvec_matrix = cvec.fit_transform(tweets['clean_tweets'])

cvec_term_df = pd.DataFrame(cvec_matrix.toarray(), columns=cvec.get_feature_names())
cvec_term_df.head()

tfidf = TfidfVectorizer(stop_words=None , ngram_range=(1, 2))

tfidf_matrix = tfidf.fit_transform(tweets['clean_tweets'])

tfidf_term_df = pd.DataFrame(tfidf_matrix.toarray(), columns=tfidf.get_feature_names())
tfidf_term_df.head()

tfidf_term_df.sum().sort_values(ascending=False).head(20)

tfidf_term_df.drop(columns=['co', 'https', 'https co', 'rt'], inplace=True)

tfidf_term_df.sum().sort_values(ascending=False).head(10) / tfidf_term_df.shape[0] * 100

cvec_term_df.drop(columns=['co', 'https', 'https co', 'rt'], inplace=True)

cvec_term_df.sum().sort_values(ascending=False).head(60) / cvec_term_df.shape[0] * 100

# https://github.com/TungPhung/Twitter-Natural-Disaster-Mapping/blob/master/ProcessingandModeling.ipynb

dbscan_cvec = DBSCAN()

important_words = ['clorox', 'alcohol', 'isopropyl', 'disinfectant', 'face mask', 'isopropyl alcohol',
                   'toilet paper', 'hand', 'sanitizer', 'hand sanitizer', 'lysol', 'handsanitizer',
                   'wipes', 'face', 'mask', 'bleach', 'masks', 'new', 'get', 'need', 'toilet', 'paper',
                   'today', 'go' , 'toiletpaper', 'spray', 'stock', 'oz', 'us', 'available']

dbscan_cvec.fit(cvec_term_df)

# checking the silhouette score
silhouette_score(cvec_term_df, dbscan_cvec.labels_)

dbscan_cvec.fit(cvec_term_df[important_words])

dbscan_cvec.labels_

tweets['commodity_tweet'] = dbscan_cvec.labels_

# checking the silhouette score
silhouette_score(cvec_term_df[important_words], dbscan_cvec.labels_)

tweets_dbscan = tweets.groupby(['tweets'])['commodity_tweet'].mean()

pd.DataFrame(tweets_dbscan).sort_values('commodity_tweet', ascending=True).head(40)

geo_dbscan = tweets.groupby(['geo'])['commodity_tweet'].mean()
pd.DataFrame(geo_dbscan).sort_values('commodity_tweet', ascending=True).head(40)

ss = StandardScaler()
tfidf_transformed = ss.fit_transform(tfidf_term_df)

tfidf_trans = DBSCAN()
tfidf_trans.fit(tfidf_term_df[important_words])

# checking the silhouette score
silhouette_score(tfidf_term_df, tfidf_trans.labels_)

tfidf_trans.labels_

tweets['source'].value_counts(normalize=True)

.333279 + .257240

tweets.groupby(['geo'])['commodity_tweet'].mean().sort_values()
