==> Code/04_All_commodities.ipynb <==
# Code for Filtering NaN and Extracting GPS Location

### Imports

import pandas as pd
import numpy as np

### Reading tweets

df = pd.read_csv('./all_commodities_tweets.csv')

df.shape

df.head()

### Filtering Missing Data from DataFrame

geo = df.loc[~df['place'].isna(), :]

geo.shape

geo.iloc[37, 8].

str = "Place(_api=<tweepy.api.API object at 0x11430af10>, id='fee8b88f74a9763d', url='https://api.twitter.com/1.1/geo/id/fee8b88f74a9763d.json', place_type='city', name='Nogales', full_name='Nogales, Sonora', country_code='MX', country='Mexico', contained_within=[], bounding_box=BoundingBox(_api=<tweepy.api.API object at 0x11430af10>, type='Polygon', coordinates=[[[-111.3534711, 31.0072089], [-110.6493311, 31.0072089], [-110.6493311, 31.377975], [-111.3534711, 31.377975]]]), attributes={})"

### Function to Extract GPS Coordinates from 'place' Column

# function to extract gps coordinates from the 'place' column. 
# returns latitude and longitude coordinates 

def get_gps(str):
    gps_str = []
    gps = []
    lat, long = [0, 0]
    place1 = str.split('coordinates=')[1][3:].split(']]]), attributes={})')[0].split('], [')
    for i in place1:
        gps_str.append(i.split(', '))
#    print(gps_str)

    for j in range(len(gps_str)):
        for k in range(2):
            gps.append(float(gps_str[j][k]))
    
#    print(gps)
    for n in range(0, 4):
        lat += gps[2*n+1] 
        long += gps[2*n]
        
    return [lat/4 , long/4]

get_gps(str)

### Mapping Function to 'place' Column

geo['gps'] = geo['place'].map(get_gps)

geo.tail()

### Write DataFrame to csv File

geo.to_csv('./tweets_gps.csv')
