==> Code/05_gmap-location.ipynb <==
# COVID-19 Commodities on Google Map 

### Imports

import gmaps
import gmaps.datasets
import pandas as pd



### Google Map API key

gmaps.configure(api_key = 'google_api_key_here')


### Commodity and GPS Location

covid_commodity = [
    {'commodity': 'clorox wipes', 'location': (44.96629517, -93.07325315), 'note': 'Kendalls Hdwr'},
    {'commodity': 'lysol & wipes', 'location': (43.5154, -79.62653), 'note': 'Clarkson Pharma'},
    {'commodity': 'masks', 'location': (32.8695093, -96.6797152), 'note': 'Leather Maters'},
    {'commodity': 'sanitizer', 'location': (32.8695093, -96.6797152), 'note': 'Rolling spokes'},
    {'commodity': 'IPA, H2O2, gloves', 'location': (35.91572, -79.06188), 'note': '501 pharmacy'},
    {'commodity': 'sanitizer', 'location': (46.50105096, -84.34907913), 'note': 'Das Gift House'},
    {'commodity': 'sanitizer', 'location': (43.79039901, -80.14427771), 'note': 'Chef Pam Joy'},
    {'commodity': 'sanitizer', 'location': (36.140987, -86.772124), 'note': 'Nashville Craft'},
    {'commodity': 'sanitizer', 'location': (42.1296, -87.8311), 'note': 'Teddie Kossof Salon Spa'},
    {'commodity': 'clorox', 'location': (35.0511565, -85.6600176), 'note': ''},
    {'commodity': 'clorox & lysol', 'location': (34.0207895, -118.4119065), 'note': 'Ace'},
    {'commodity': 'lysol', 'location': (43.629311, -79.2725695), 'note': 'Walmart'},
    {'commodity': 'lysol & clorox', 'location': (43.555244, -79.616073), 'note': 'Clarkson Pharma'},
    {'commodity': 'toilet paper', 'location': (36.20000805, -86.791549), 'note': ''},
    {'commodity': 'toilet paper', 'location': (36.20000805, -86.791549), 'note': ''},
    {'commodity': 'toilet paper', 'location': (34.19999172983301, -118.37995234955565), 'note': 'Ace'},
    {'commodity': 'sanitizer', 'location': (40.5833894, -75.5588775), 'note': ''},
    {'commodity': 'sanitizer', 'location': (26.275662, -80.096419), 'note': 'Zakia Miami'},
    {'commodity': 'sanitizer', 'location': (45.424871, -122.79185375), 'note': 'Rose City Distilling'},
    {'commodity': 'sanitizer', 'location': (35.2264115, -81.0378265), 'note': 'Carolina Rum'},
    {'commodity': 'sanitizer & mask', 'location': (33.7671944, -84.433106), 'note': ''},
    {'commodity': 'sanitizer', 'location': (34.22887575, -118.235241), 'note': 'Ralphs Grocery'},
    {'commodity': 'sanitizer', 'location': (40.787501, -74.1271651), 'note': 'Gaspers Best'},
    {'commodity': 'sanitizer', 'location': (45.53640175, -122.63090815), 'note': 'Long D. Pharma'},
    {'commodity': 'sanitizer', 'location': (42.24571305, -83.6018025), 'note': 'Butters Hygienics'},
    {'commodity': 'sanitizer', 'location': (45.53640175, -122.63090815), 'note': 'Long D. Pharma'},
    {'commodity': 'toilet paper', 'location': (47.2315974, -123.1239325), 'note': ''},
    {'commodity': 'toilet paper', 'location': (42.308346, 0.3142995), 'note': 'AM Cleaning'},
    {'commodity': 'toilet paper', 'location': (29.8384948, -95.4464865), 'note': ''},
    {'commodity': 'toilet paper', 'location': (35.1170026, -106.6327181), 'note': 'Costco'},
    {'commodity': 'toilet paper', 'location': (33.75133608416934, -84.32186124438131), 'note': 'ATL Vision'},
    {'commodity': 'toilet paper', 'location': (39.1018362, -77.5530482), 'note': 'CS Hubscher'},
    {'commodity': 'toilet paper', 'location': (41.144642, -73.3425245), 'note': ''},
    {'commodity': 'toilet paper', 'location': (30.199115109821854, -81.82494668408846), 'note': ''},
    {'commodity': 'toilet paper', 'location': (25.6719858, -80.34711985), 'note': ''}
    
]

comm_locations = [comm['location'] for comm in covid_commodity]
info_box_template = """
<dl>
<dt>Commodity</dt><dd>{commodity}</dd>
<dt>Note</dt><dd>{note}</dd>
</dl>
"""
comm_info = [info_box_template.format(**i) for i in covid_commodity]

marker_layer = gmaps.marker_layer(comm_locations, info_box_content=comm_info)
fig = gmaps.figure()
fig.add_layer(marker_layer)
fig
